requirements.txt
Django>=4.2,<5.0
djangorestframework
django-cors-headers
gunicorn
psycopg2-binary
pillow
python-dotenv
